requirements.txt
PyGithub>=1.58.0
yamllint>=1.31.0
ruamel.yaml
ruamel.yaml.clib
